pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kabolai"
version = "0.1.0"
description = "KA-BOL-AI: Bilingual voice-controlled PC assistant (English + Ukrainian)"
requires-python = ">=3.9"
license = "MIT"

dependencies = [
    # Core
    "pyyaml>=6.0",
    "click>=8.1.0",
    "pydantic>=2.0",
    "requests>=2.31.0",

    # Audio
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",

    # STT (CPU default)
    "vosk>=0.3.45",

    # TTS English (CPU fallback)
    "pyttsx3>=2.90",

    # TTS Ukrainian
    "ukrainian-tts @ git+https://github.com/robinhad/ukrainian-tts.git",

    # UI
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "keyboard>=0.13.5",

    # System info
    "psutil>=5.9.0",
]

[project.optional-dependencies]
gpu = [
    "faster-whisper>=1.0.0",
    "piper-tts>=1.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
kabolai = "kabolai.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
